Incorporate PR review feedback
